pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devcontainer-manager"
version = "0.1.0"
description = "Interactive tooling for managing devcontainer.json features"
authors = [{name = "OpsContainers", email = "oss@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "rich>=13",
  "pyyaml>=6",
]

[project.scripts]
devcontainer-manager = "devcontainer_manager.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["devcontainer_manager"]
